[nom] Added in-memory cache for feeds preview
